Write a little test to prove that the conductor can handle a shared_ptr without causing a memory exception.
This works, although interesting requires the use of Run rather than Start which implies it is running on the main thread.

Change the STUN servers to something that might actually work on campus.
It would probably be a good idea to set  this to a long list of potential servers that might work.